setImageResource won't set an image if a default drawable already exists
When updateDrawable(null) is called, the default drawable is set as
mContentDrawable. Later when we attempt to set the image we try to
confirm that there is no content drawable before proceeding which fails
because the default is now the content. To correct this, I made an
addition to check that mContentDrawable is not mDefaultDrawable.